Refactor LSP server related logic into Server and keep the Service as a thin abstraction
