+++ server/aws-lsp-codewhisperer/src/client/sigv4/codewhisperer.ts
@@ -3,7 +3,7 @@ import { ServiceConfigurationOptions } from 'aws-sdk/lib/service'
 const apiConfig = require('./service-2.json')
 import CodeWhispererClient = require('./codewhispererclient')
 
-export function createCodeWhispererClient(options: ServiceConfigurationOptions): CodeWhispererClient {
+export function createCodeWhispererSigv4Client(options: ServiceConfigurationOptions): CodeWhispererClient {
     return createService(options) as CodeWhispererClient
 }
