+++ server/aws-lsp-codewhisperer/src/client/sigv4/codewhispererclient.d.ts
