+++ server/aws-lsp-codewhisperer/src/client/sigv4/service-2.json
