+++ server/aws-lsp-codewhisperer/src/language-server/codeWhispererServer.ts
@@ -1,14 +1,85 @@
 import { Auth } from '@aws-placeholder/aws-language-server-runtimes/out/features'
 import { InlineCompletionParams } from '@aws-placeholder/aws-language-server-runtimes/out/features/lsp/inline-completions/futureProtocol'
-import { InlineCompletionList } from '@aws-placeholder/aws-language-server-runtimes/out/features/lsp/inline-completions/futureTypes'
+import { InlineCompletionContext, InlineCompletionItem, InlineCompletionList, InlineCompletionTriggerKind } from '@aws-placeholder/aws-language-server-runtimes/out/features/lsp/inline-completions/futureTypes'
 import { Server } from '@aws-placeholder/aws-language-server-runtimes/out/runtimes'
 import { CancellationToken } from 'vscode-languageserver'
-import { CodeWhispererServiceBase, CodeWhispererServiceIAM, CodeWhispererServiceToken } from './codeWhispererService'
+import { Position, TextDocument } from 'vscode-languageserver-textdocument'
+import { CodeWhispererServiceBase, CodeWhispererServiceIAM, CodeWhispererServiceToken, GenerateSuggestionsRequest, Suggestion } from './codeWhispererService'
 import { getSupportedLanguageId } from './languageDetection'
+import { truncateOverlapWithRightContext } from './utils'
+
+interface DoInlineCompletionParams {
+    textDocument: TextDocument
+    position: Position
+    context: InlineCompletionContext
+    token?: CancellationToken
+    inferredLanguageId: string
+}
+
+interface GetSuggestionsParams {
+    textDocument: TextDocument
+    position: Position
+    maxResults: number
+    token: CancellationToken
+    inferredLanguageId: string
+}
+
 
 export const CodewhispererServerFactory = (service: (auth: Auth) => CodeWhispererServiceBase): Server =>
     ({ auth, lsp, workspace, logging }) => {
         const codeWhispererService = service(auth)
+
+        // const setCustomisation= async (a: any) : any => {
+        //     // call the underlying codewhisperer service, or set a flag to 
+        // }
+
+        const getSuggestions = async (params: GetSuggestionsParams): Promise<Suggestion[]> => {
+            const left = params.textDocument.getText({
+                start: { line: 0, character: 0 },
+                end: params.position,
+            })
+            const right = params.textDocument.getText({
+                start: params.position,
+                end: params.textDocument.positionAt(params.textDocument.getText().length),
+            })
+
+            const request: GenerateSuggestionsRequest = {
+                fileContext: {
+                    filename: params.textDocument.uri,
+                    programmingLanguage: {
+                        languageName: params.inferredLanguageId,
+                    },
+                    leftFileContent: left,
+                    rightFileContent: right,
+                },
+                maxResults: params.maxResults,
+            }
+
+            return codeWhispererService.generateSuggestions(request);
+        }
+
+        const doInlineCompletion = async (params: DoInlineCompletionParams): Promise<InlineCompletionList | null> => {
+            const recommendations = await getSuggestions({
+                textDocument: params.textDocument,
+                position: params.position,
+                maxResults: params.context.triggerKind == InlineCompletionTriggerKind.Automatic ? 1 : 5,
+                token: params.token || CancellationToken.None,
+                inferredLanguageId: params.inferredLanguageId
+            })
+
+            const items: InlineCompletionItem[] = recommendations.map<InlineCompletionItem>(r => {
+                return {
+                    insertText: truncateOverlapWithRightContext(params.textDocument, r.content, params.position),
+                    range: params.context.selectedCompletionInfo?.range,
+                }
+            })
+
+            const completions: InlineCompletionList = {
+                items,
+            }
+
+            return completions
+        }
 
         const onInlineCompletionHandler = async (
             params: InlineCompletionParams,
@@ -19,10 +90,10 @@ export const CodewhispererServerFactory = (service: (auth: Auth) => CodeWhispere
             }
 
             const textDocument = await workspace.getTextDocument(params.textDocument.uri)
-            const languageId = getSupportedLanguageId(textDocument)
+            const languageId = getSupportedLanguageId(textDocument)"
             if (textDocument && languageId) {
                 try {
-                    const recommendations = await codeWhispererService.doInlineCompletion({
+                    const recommendations = await doInlineCompletion({
                         textDocument,
                         position: params.position,
                         context: { triggerKind: params.context.triggerKind },
